refactor(api): PR3 — extract telegram service to api/telegram.py
Move build_telegram_message, push_telegram_direct (DEPRECATED #162 shim
that delegates to notifier.notify(SignalEvent(...))), _send_telegram_raw,
push_webhook from btc_api.py to api/telegram.py. Plus the _TELEGRAM_API
constant.

api/telegram.py is a service module — NO APIRouter, telegram is
outbound-only. Signal-filter logic (should_notify_signal,
_is_duplicate_signal, _mark_notified) stays in btc_api.py until PR5
where it joins the signals domain.

The lazy `from health import get_symbol_state` inside push_telegram_direct
is preserved — health.py has its own deps that could cycle.

btc_api.py re-exports all 4 functions for legacy callers (scanner_loop)
until PR7.

New unit tests cover: message format for active signals + neutral state,
_send_telegram_raw URL/auth + skip on missing creds, push_webhook skip
on missing URL, push_webhook DB audit row insertion.

test_api.py: update test_notifies_on_premium_signal to patch
api.telegram.notify (the actual call site after PR3 extraction) instead
of btc_api.notify.

1008 tests pass.
